LPS-129801 Port all hooks to TypeScript
The goal here is to get rid of _most_ of the ugly `@ts-ignore`
suppressions in the index file. Only `render.es` remains, and the reason
I didn't port that one yet is because it is going to require more
extensive changes:

- Need `@types/react-dom` in the dependency graph (technically, should
  come in via `@liferay/npm-scripts`, so that will require a separate
  PR).
- Need to tweak build to use `"tsx": "react"`; for some reason, I am
  getting "TS17004: Cannot use JSX unless the '--jsx' flag is provided."
  even when I provide the flag on the command-line (ie. bypassing
  `npm-scripts`, just calling `tsc --noEmit --jsx react`); it also fails
  when added to the `tsconfig.json`. Not sure what is going on there
  because the option is being picked up in some way (if you spell the
  option wrong, it complains; if you omit the option, you get more
  errors; the problem is that turning it on only suppresses _some_ of
  the errors.... weird)

Also didn't port the tests, but it obviously makes sense to do those too
when I have more time.
